header.php - bug fix javascript error href attr #
In Console: Uncaught Error: Syntax error, unrecognized expression: #
Since bootstrap.js uses the href attr for the selector and # is an invalid javascript selector, I changed the href attr value to "menu". It doesn't matter what it is as long as it's a valid JavaScript selector, since the link's default behavior is prevented.